refactor: add ui-logic directory
