+++ src/view/ui-logic/rendering.ts
@@ -8,7 +8,7 @@ import {
   TEStartTag
 } from '@/parser/template/types'
 import { DefaultValue } from '@/parser/script/types'
-import { evalWithScope } from './eval'
+import { evalWithScope } from '../eval'
 import { isObject, range, mapValues } from '@/utils'
 
 function attributeValue(
